+++ lib/project_types/extension/models/specification_handlers/theme_app_extension.rb
@@ -75,7 +75,7 @@ module Extension
 
         def serve(**options)
           @ctx = options[:context]
-          port = options[:port]
+          # port = options[:port]
           root = options[:context]&.root
 
           ShopifyCLI::Theme::DevServer::AppExtensionDevServer.start(@ctx, root)
